Extract register, execute function.
