+++ uniswap-smartwallet/README.md
@@ -1,9 +1,17 @@
-# Testing for Smart Wallet
+# Testing for Unitswap Calibur contract.
 
-1. Register a key and make a transfer WETH with Calibur.
+1. Register a key and make a transfer WETH with alternative signer account.
 
-	```
-	export EOA_KEY=0xEoaKey
-	export RELAY_KEY=0xRelayKey
-	yarn run transfer
-	```
+  ```
+  export EOA_KEY=0xEoaKey
+  export SIGNER_KEY=0xSignerKey
+  yarn run transfer
+  ```
+
+2. Swap ETH for UNI using an alternative signer account.
+
+  ```
+  export EOA_KEY=0xEoaKey
+  export SIGNER_KEY=0xSignerKey
+  yarn run swap
+  ```
